perf(testing): consolidate integration_docker classes onto shared stacks
Every test class boots and tears down its own 12-container compose stack
(~40s of fixture overhead per class). Four merges remove four boots while
keeping every assertion scoped to its own nodes:

- The node-event and group-event related-resources classes share one
  stack: disjoint kinds, both schemas loaded in a single converged call,
  and every event count is filtered by primary_node__ids.
- The two repository classes share one stack; the operational-status
  assertion now fetches its repository by name so the tests stay
  order-independent.
- The triggered-actions test joins TestProposeChangeRepository, which
  already builds the identical car-dealership dataset (same schema, same
  "people" group, same repository); the duplicated schema+data+repo-sync
  load goes away. The trigger test stays last so no trigger rule exists
  while earlier tests mutate persons.
- The display-label backfill tests join the file-object class: disjoint
  kinds, id-scoped assertions.

Full suite green at -n 4 loadscope: 280s (was 358s with 13 stacks;
serial baseline 1198s). 63 passed, 1 skipped, 1 xfailed - one duplicated
schema-load test was absorbed by the events merge.
